+++ dart/lib/src/transport/http_transport.dart
@@ -12,7 +12,7 @@ import '../sentry_envelope.dart';
 import 'transport.dart';
 import 'rate_limiter.dart';
 import '../http_client/client_provider.dart'
-    if (dart.library.io) '../http_client/io_proxy_client_provider.dart';
+    if (dart.library.io) '../http_client/io_client_provider.dart';
 
 /// A transport is in charge of sending the event to the Sentry server.
 class HttpTransport implements Transport {
